+++ lua/plugins/which-key.lua
@@ -1,0 +1,62 @@
+return {
+  "folke/which-key.nvim",
+  event = "VeryLazy",
+  config = function()
+    require("config.keymaps").setup()
+  end,
+  -- opts = {
+  --   plugins = {
+  --     marks = false,     -- shows a list of your marks on ' and `
+  --     registers = true, -- shows your registers on " in NORMAL or <C-r> in INSERT mode
+  --     -- the presets plugin, adds help for a bunch of default keybindings in Neovim
+  --     -- No actual key bindings are created
+  --     spelling = {
+  --       enabled = true,   -- enabling this will show WhichKey when pressing z= to select spelling suggestions
+  --       suggestions = 20, -- how many suggestions should be shown in the list?
+  --     },
+  --     presets = {
+  --       operators = true,    -- adds help for operators like d, y, ...
+  --       motions = true,      -- adds help for motions
+  --       text_objects = true, -- help for text objects triggered after entering an operator
+  --       windows = true,      -- default bindings on <c-w>
+  --       nav = true,          -- misc bindings to work with windows
+  --       z = true,            -- bindings for folds, spelling and others prefixed with z
+  --       g = true,            -- bindings for prefixed with g
+  --     },
+  --   },
+  --   operators = { gc = "Comments" },
+  --   key_labels = {
+  --     -- override the label used to display some keys. It doesn't effect WK in any other way.
+  --     -- For example:
+  --     -- ["<space>"] = "SPC",
+  --     -- ["<cr>"] = "RET",
+  --     -- ["<tab>"] = "TAB",
+  --   },
+  --   motions = {
+  --     count = true,
+  --   },
+  --   icons = {
+  --     breadcrumb = "»", -- symbol used in the command line area that shows your active key combo
+  --     separator = "➜", -- symbol used between a key and it's label
+  --     group = "+", -- symbol prepended to a group
+  --   },
+  --   popup_mappings = {
+  --     scroll_down = "<C-d>", -- binding to scroll down inside the popup
+  --     scroll_up = "<C-u>",   -- binding to scroll up inside the popup
+  --   },
+  --   window = {
+  --     border = "single",          -- none, single, double, shadow
+  --     position = "bottom",      -- bottom, top
+  --     margin = { 1, 2, 1, 2 },  -- extra window margin [top, right, bottom, left]. When between 0 and 1, will be treated as a percentage of the screen size.
+  --     padding = { 1, 2, 1, 2 }, -- extra window padding [top, right, bottom, left]
+  --     winblend = 0,             -- value between 0-100 0 for fully opaque and 100 for fully transparent
+  --     zindex = 1000,            -- positive value to position WhichKey above other floating windows.
+  --   },
+  --   layout = {
+  --     height = { min = 4, max = 25 }, -- min and max height of the columns
+  --     width = { min = 20, max = 50 }, -- min and max width of the columns
+  --     spacing = 3, -- spacing between columns
+  --     align = "left", -- align columns left, center or right
+  --   },
+  -- }
+}
